Reorganize files and change behavior of pretty format
